% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclide.R
\name{Euclide_R}
\alias{Euclide_R}
\title{Computing euclidian distance}
\usage{
Euclide_R(v, w)
}
\arguments{
\item{v}{a vector of numeric data}

\item{w}{a vector of numeric data}
}
\value{
the euclidian distance between v and w
}
\description{
Euclidian Distance with R functions
}
